Moved advanced details to separate python file
